完善升级的流程&排除一些bug

+++ shellscript/update_firmware.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
 #update firmware
-
 DOWNLOADDIR="/home/firmware"
 UPDATESTATUS="/home/ubuntu/update_status"
 #FIRMWARE=$(ls /home/download/ | grep firmware)
@@ -28,28 +27,12 @@ then
 	apt-get install rsync
 fi
 
-kill_little_watchdog() {
-	pid=$(ps -ef | grep little_feed | grep -v grep | awk -F" " '{print $2}')
-	if [ -n "${pid}" ]
-	then 
-		kill -9 ${pid}
-	fi
-}
-
 update_start() {
-	#1.close monitor script and mqtt client process
-	#  start_network.sh must keep running,beacuse feed watchdog in this process
-	echo "start update"
-	#/root/watch.sh zywlmonitor.sh stop
-	#/root/watch.sh mqtt stop
-	systemctl stop zywldl
-	#2.show update image
 	python3 /root/showimage.py 4
 	echo "tarfile:${version}:0" > ${UPDATESTATUS}
 }
 
 tarfile() {
-	echo "tar file"
 	if [ -d "${DOWNLOADDIR}.back" ]
 	then 
 		rm -rf "${DOWNLOADDIR}.back"
@@ -62,19 +45,6 @@ tarfile() {
 
 	mkdir -p ${DOWNLOADDIR}
 
-	sync
-	while true
-	do
-		dir=$(pwd)
-		if [ "${dir}" != "${DOWNLOADDIR}" ]
-		then
-			mkdir -p ${DOWNLOADDIR}
-			cd ${DOWNLOADDIR}
-		else
-			break
-		fi
-	done
-	
 	#该文件必须首先要存在/root目录下才能解压
 	python3 /root/uncompress.py ${FIRMWARE} ${DOWNLOADDIR}
 	if [ ! -d "${DOWNLOADDIR}/target" ]
@@ -85,14 +55,16 @@ tarfile() {
 	fi
 
 	sleep 1
-	rsync ${DOWNLOADDIR}/target/shell/little_feed.sh /home/watchdog
+	#rsync ${DOWNLOADDIR}/target/shell/little_feed.sh /home/watchdog
 	#rsync python feed watchdog script to /home/watchdog
-	rsync -r ${DOWNLOADDIR}/target/watchdog/ /home/watchdog
+	if [ -d "${DOWNLOADDIR}/target/watchdog" ];then
+		rsync -r ${DOWNLOADDIR}/target/watchdog/ /home/watchdog
+	fi
+
 	echo "move:${version}:0" > ${UPDATESTATUS}
 }
 
 move() {
-	echo "move file"
 	cd /root
 	if [ ! -d "${DOWNLOADDIR}/target" ]
 	then 
@@ -100,74 +72,95 @@ move() {
 		echo "error:${version}:3" > ${UPDATESTATUS}
 	else
 		#删除只存在于/root中，不存在于更新包中的文件
-		#/root/watch.sh stop
-		systemctl stop zywldl
 		sleep 1
-		/home/watchdog/little_feed.sh &
-		
-		#rsync python modules
-		rsync -r ${DOWNLOADDIR}/target/python/*.so /usr/local/lib/python3.5/dist-packages/
 
 		#rsync shell script to /root
-		rsync -r ${DOWNLOADDIR}/target/shell/*.sh /root/
-		
+		if [ -d "${DOWNLOADDIR}/target/shell" ];then
+			if [ -f "${DOWNLOADDIR}/target/shell/update_firmware.sh" ];then
+				diff ${DOWNLOADDIR}/target/shell/update_firmware.sh /root/update_firmware.sh > /dev/null
+				if [ $? -eq 1 ];then
+					rsync ${DOWNLOADDIR}/target/shell/update_firmware.sh /root/update_firmware.sh
+					exit
+				fi
+			fi		
+			rsync -r ${DOWNLOADDIR}/target/shell/*.sh /root/
+		fi
+				
+		#rsync python modules
+		if [ -d "${DOWNLOADDIR}/target/python" ];then
+			rsync -r ${DOWNLOADDIR}/target/python/*.so /usr/local/lib/python3.5/dist-packages/
+		fi
+
 		#rsync python run script to /root
-		rsync -r ${DOWNLOADDIR}/target/run/*.py /root/
-		
+		if [ -d "${DOWNLOADDIR}/target/run" ];then
+			rsync -r ${DOWNLOADDIR}/target/run/*.py /root/
+		fi
+
 		#rsync image file to /root
-		rsync -r ${DOWNLOADDIR}/target/image /root/
-		
+		if [ -d "${DOWNLOADDIR}/target/image" ];then
+			rsync -r ${DOWNLOADDIR}/target/image /root/
+		fi
+
 		#rsync image file to /root
-		rsync -r ${DOWNLOADDIR}/target/ppp/ /etc/ppp
+		if [ -d "${DOWNLOADDIR}/target/ppp" ];then
+			rsync -r ${DOWNLOADDIR}/target/ppp /etc/ppp/
+		fi
 
 		#config file rsync
-		if [ ! -f "/root/net.conf" ]
-		then
+		if [ ! -f "/root/net.conf" ];then
 			cp ${DOWNLOADDIR}/target/net.conf /root
 		fi
 
-		if [ ! -f "/root/config.ini" ]
-		then
+		if [ ! -f "/root/config.ini" ];then
 			cp ${DOWNLOADDIR}/target/config.ini /root
 		fi
 
-		rsync -r ${DOWNLOADDIR}/target/ntp.conf /etc/
-		
-		#crtfile rsync
-		mkdir -p /root/crtfile
-		rsync -r ${DOWNLOADDIR}/target/*.crt /root/crtfile/
+		if [ -f "${DOWNLOADDIR}/target/ntp.conf " ];then
+			rsync -r ${DOWNLOADDIR}/target/ntp.conf /etc/
+		fi
 
-		if [ ! -d "/root/log" ]
-		then
+		#crtfile rsync
+		if [ -f "${DOWNLOADDIR}/target/mqtt.iotwonderful.cn.crt" ];then
+			mkdir -p /root/crtfile
+			rsync -r ${DOWNLOADDIR}/target/mqtt.iotwonderful.cn.crt /root/crtfile/
+		fi
+
+		if [ ! -d "/root/log" ];then
 			mkdir -p /root/log
 		fi
 
 		#install frpc service
-		rsync -r ${DOWNLOADDIR}/target/frp/systemd/* /lib/systemd/system/
-		rsync -r ${DOWNLOADDIR}/target/frp/frpc /usr/bin/
-		rsync -r ${DOWNLOADDIR}/target/frp/frpc.ini /etc/frp/
-		systemctl enable frpc
+		if [ -d "${DOWNLOADDIR}/target/frp" ];then
+			if [ -d "${DOWNLOADDIR}/target/frp/systemd" ];then
+				rsync -r ${DOWNLOADDIR}/target/frp/systemd/* /lib/systemd/system/
+			fi
+			if [ -f "${DOWNLOADDIR}/target/frp/frpc" ];then
+				rsync -r ${DOWNLOADDIR}/target/frp/frpc /usr/bin/
+			fi
+			if [ -f "${DOWNLOADDIR}/target/frp/frpc.ini" ];then
+				rsync -r ${DOWNLOADDIR}/target/frp/frpc.ini /etc/frp/
+			fi
+		fi
 
 		#ver=${FIRMWARE#*_}
 		#ver=${ver%%.*}
 		/root/set_config.sh "version " " ${version}"
 		/root/set_config.sh "VERSION" "${version}"
 
-		rsync ${DOWNLOADDIR}/target/rc.local /etc/
+		#rsync ${DOWNLOADDIR}/target/rc.local /etc/
+		if [ -f "${DOWNLOADDIR}/systemd/zywldl.service" ];then
+			rsync ${DOWNLOADDIR}/systemd/zywldl.service /lib/systemd/system/
+			systemctl enable zywldl
+		fi
 
 		echo "clear:${version}:0" > ${UPDATESTATUS}
 	fi
 }
 
 update_clear() {
-	echo "clear file"
 	cd /root
 	rm /home/download/*
 	echo "done:${version}:0" > ${UPDATESTATUS}
-	sync
-	kill_little_watchdog
-	sleep 1
-	#/root/watch.sh restart
 	systemctl restart zywldl
 }
